pyproject.toml
[project]
name = "Game-ClientServer"
description = "Tool to collect info about dependencies in pivot projects from GitLab repositories and visualize them via tree graph"
version = "0.0.1"
requires-python = ">= 3.9"

authors = [
  {name = "Artem Ansimov", email="artansimov@yandex.ru"}
]

## Dependencies

dependencies = [
  "pydantic==2.7.4",
  "cryptography"
]

[project.optional-dependencies]
dev = [
    "python_lsp_server==1.11.0",
    "python_lsp_server[pylint]",
    "pylint",
    "black==23.3.0",
]
server = [
  "sqlalchemy[aiomysql]",
]
test = [
  "pytest-asyncio",
  "PyHamcrest"
]


## Pytest

[tool.pytest.ini_options]
log_cli = true


## Pylint

[tool.pylint.basic]
good-names = ["f", "_"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint."MESSAGES CONTROL"]

disable = '''
  missing-class-docstring,
  missing-module-docstring,
  missing-function-docstring
  '''

## Black

[tool.black]
target-version = ['py39']
line-length = 120

